// src/common/enums/role.enum.ts
export enum DefaultRole {
  SUPER_ADMIN = 'SUPER_ADMIN',
  PRINCIPAL = 'PRINCIPAL',
  SCHOOL_ADMIN = 'SCHOOL_ADMIN',
  ACCOUNTANT = 'ACCOUNTANT',
  HR_MANAGER = 'HR_MANAGER',
  TEACHER = 'TEACHER',
  CLASS_TEACHER = 'CLASS_TEACHER',
  RECEPTIONIST = 'RECEPTIONIST',
  FEE_COLLECTOR = 'FEE_COLLECTOR',
  EXAM_COORDINATOR = 'EXAM_COORDINATOR',
  STAFF = 'STAFF',
  LIBRARIAN = 'LIBRARIAN',
  TRANSPORT_MANAGER = 'TRANSPORT_MANAGER',
  PARENT = 'PARENT',
  STUDENT = 'STUDENT',
}

export enum UserType {
  ADMIN = 'ADMIN',
  STAFF = 'STAFF',
  TEACHER = 'TEACHER',
  ACCOUNTANT = 'ACCOUNTANT',
  PARENT = 'PARENT',
  STUDENT = 'STUDENT',
}